% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cores.R
\name{cores_insper}
\alias{cores_insper}
\title{Cores do insper}
\usage{
cores_insper(n_cores = 5)
}
\arguments{
\item{n_cores}{quantidade de cores (até 15)}
}
\value{
vetor de cores
}
\description{
Cores do insper
}
