% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_cores}
\alias{da_cores}
\title{da_cores}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 4 columns.
}
\usage{
da_cores
}
\description{
da_cores
}
\keyword{datasets}
